requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Database & Caching
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# HTTP & WebSockets
requests>=2.31.0
aiohttp>=3.9.0
websockets>=12.0
httpx>=0.25.0

# AI & ML Dependencies
torch>=2.6.0
torchvision>=0.19.0
torchaudio>=2.6.0
transformers>=4.36.0
sentence-transformers>=2.2.0
clip-by-openai>=1.0

# Vector Database
qdrant-client>=1.7.0

# Image & Video Processing
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Audio Processing
gtts>=2.4.0
pydub>=0.25.0

# Performance & Monitoring
psutil>=5.9.0
nvidia-ml-py3>=7.352.0